Move sweettest into holochain crate root
