+++ crates/holochain/src/core/ribosome/host_fn/capability_grants.rs
@@ -17,10 +17,8 @@ pub fn capability_grants(
 #[cfg(feature = "slow_tests")]
 pub mod wasm_test {
     use crate::fixt::ZomeCallHostAccessFixturator;
-    use crate::{conductor::ConductorBuilder, test_utils::sweettest::SweetConductor};
-    use crate::{
-        core::workflow::call_zome_workflow::CallZomeWorkspace, test_utils::sweettest::SweetDnaFile,
-    };
+    use crate::{conductor::ConductorBuilder, sweettest::SweetConductor};
+    use crate::{core::workflow::call_zome_workflow::CallZomeWorkspace, sweettest::SweetDnaFile};
     use ::fixt::prelude::*;
     use hdk::prelude::*;
     use holochain_types::fixt::CapSecretFixturator;
