+++ crates/holochain/src/core/workflow/validation_receipt_workflow/tests.rs
@@ -1,5 +1,5 @@
+use crate::sweettest::*;
 use crate::test_utils::consistency_10s;
-use crate::test_utils::sweettest::*;
 use fallible_iterator::FallibleIterator;
 use hdk::prelude::*;
 use holo_hash::DhtOpHash;
