+++ crates/holochain/src/lib.rs
@@ -10,6 +10,10 @@ pub mod conductor;
 pub mod core;
 #[allow(missing_docs)]
 pub mod fixt;
+
+#[cfg(any(test, feature = "test_utils"))]
+#[deny(missing_docs)]
+pub mod sweettest;
 #[cfg(any(test, feature = "test_utils"))]
 #[deny(missing_docs)]
 pub mod test_utils;
